Fix doctests with dev numpy

+++ ndindex/tuple.py
@@ -246,8 +246,8 @@ class Tuple(NDIndex):
         Integer(1)
         >>> a[..., 1]
         array(1)
-        >>> a[1]
-        1
+        >>> a[1] # doctest: +SKIP38
+        np.int64(1)
 
         See https://github.com/Quansight-Labs/ndindex/issues/22.
